+++ src/output/p_flac.cpp
@@ -83,24 +83,14 @@ flac_packetizer_c::process(packet_cptr packet) {
 }
 
 connection_result_e
-flac_packetizer_c::can_connect_to(generic_packetizer_c *src,
-                                  std::string &error_message) {
-  flac_packetizer_c *fsrc = dynamic_cast<flac_packetizer_c *>(src);
-  if (!fsrc)
-    return CAN_CONNECT_NO_FORMAT;
-
-  connect_check_a_samplerate(m_stream_info.sample_rate, fsrc->m_stream_info.sample_rate);
-  connect_check_a_channels(m_stream_info.channels, fsrc->m_stream_info.channels);
-  connect_check_a_bitdepth(m_stream_info.bits_per_sample, fsrc->m_stream_info.bits_per_sample);
-
-  if (   (m_header->get_size() != fsrc->m_header->get_size())
-      || !m_header
-      || !fsrc->m_header
-      || memcmp(m_header->get_buffer(), fsrc->m_header->get_buffer(), m_header->get_size())) {
-    error_message = fmt::format(Y("The codec's private data does not match (lengths: {0} and {1})."), m_header->get_size(), fsrc->m_header->get_size());
-    return CAN_CONNECT_MAYBE_CODECPRIVATE;
-  }
-
-  return CAN_CONNECT_YES;
+flac_packetizer_c::can_connect_to(generic_packetizer_c */* src */,
+                                  std::string &/* error_message */) {
+  return CAN_CONNECT_NO_UNSUPPORTED;
 }
+
+split_result_e
+flac_packetizer_c::can_be_split(std::string &/* error_message */) {
+  return CAN_SPLIT_NO_UNSUPPORTED;
+}
+
 #endif